Separate email analysis component and add headers body section
